style: code style according to prettier config

+++ src/commands/moderation/nuke.ts
@@ -1,4 +1,12 @@
-import { ActionRowBuilder, AttachmentBuilder, ButtonBuilder, ButtonStyle, ChannelType, type ComponentType, type Interaction } from "discord.js";
+import {
+  ActionRowBuilder,
+  AttachmentBuilder,
+  ButtonBuilder,
+  ButtonStyle,
+  ChannelType,
+  type ComponentType,
+  type Interaction
+} from "discord.js";
 import { CommandBuilder } from "../../structures/CommandBuilder.js";
 import { CommandType } from "../../types/command.types.js";
 import { getChannel } from "../../util/functions.js";
@@ -16,7 +24,12 @@ export default new CommandBuilder({
   async run(client, message, args) {
     const channel = getChannel(args[1], message) || message.channel;
     if (channel.type !== ChannelType.GuildText)
-      return message.reply(client.beautifyMessage("Solo puedo hacer esta acción con canales de texto.", { emoji: "error" }));
+      return message.reply(
+        client.beautifyMessage(
+          "Solo puedo hacer esta acción con canales de texto.",
+          { emoji: "error" }
+        )
+      );
 
     const row = new ActionRowBuilder<ButtonBuilder>().addComponents(
       new ButtonBuilder()
@@ -27,15 +40,26 @@ export default new CommandBuilder({
         .setLabel("No")
         .setStyle(ButtonStyle.Danger)
         .setCustomId("nuke-btn-no")
-    )
+    );
 
-    const confirmMessage = await message.reply({ content: "> ¿Estás seguro de hacer esto?", components: [row] });
+    const confirmMessage = await message.reply({
+      content: "> ¿Estás seguro de hacer esto?",
+      components: [row]
+    });
     const filter = (m: Interaction) => m.user.id === message.author.id;
-    const collector = confirmMessage.createMessageComponentCollector<ComponentType.Button>({ filter, max: 1, maxUsers: 1, time: 20_000 });
+    const collector =
+      confirmMessage.createMessageComponentCollector<ComponentType.Button>({
+        filter,
+        max: 1,
+        maxUsers: 1,
+        time: 20_000
+      });
 
-    const imgAttach = new AttachmentBuilder("https://imgur.com/LIyGeCR.gif", { name: "nuke.gif" });
+    const imgAttach = new AttachmentBuilder("https://imgur.com/LIyGeCR.gif", {
+      name: "nuke.gif"
+    });
 
-    collector.once("collect", async (res) => {
+    collector.once("collect", async res => {
       if (res.customId === "nuke-btn-yes") {
         try {
           const newChannel = await channel.clone();
@@ -44,28 +68,51 @@ export default new CommandBuilder({
 
           await channel.delete();
 
-          await newChannel.send({ content: `**El canal ha explotado en mil pedazos...**`, files: [imgAttach] })
+          await newChannel.send({
+            content: `**El canal ha explotado en mil pedazos...**`,
+            files: [imgAttach]
+          });
         } catch (error) {
           console.error(error);
-          message.channel.send(client.beautifyMessage(`Ocurrió un error al intentar clonar y eliminar el canal.`, { emoji: "warning" }));
+          message.channel.send(
+            client.beautifyMessage(
+              `Ocurrió un error al intentar clonar y eliminar el canal.`,
+              { emoji: "warning" }
+            )
+          );
         }
       } else {
-        confirmMessage.edit({ content: client.beautifyMessage("Operación cancelada.", { emoji: "error" }), components: [] })
-          .then(msg => setTimeout(() => {
-            if (msg.deletable)
-              msg.delete();
-          }, 5000));
+        confirmMessage
+          .edit({
+            content: client.beautifyMessage("Operación cancelada.", {
+              emoji: "error"
+            }),
+            components: []
+          })
+          .then(msg =>
+            setTimeout(() => {
+              if (msg.deletable) msg.delete();
+            }, 5000)
+          );
       }
     });
 
     collector.once("end", (_, reason) => {
       if (reason === "limit") return;
       if (reason === "time")
-      confirmMessage.edit({ content: client.beautifyMessage(`Operación cancelada por inactividad.`, { emoji: "error" }), components: [] })
-        .then(msg => setTimeout(() => {
-          if (msg.deletable)
-          msg.delete();
-        }, 5000));
+        confirmMessage
+          .edit({
+            content: client.beautifyMessage(
+              `Operación cancelada por inactividad.`,
+              { emoji: "error" }
+            ),
+            components: []
+          })
+          .then(msg =>
+            setTimeout(() => {
+              if (msg.deletable) msg.delete();
+            }, 5000)
+          );
     });
 
     return;
